Move scheme client to a separate module
