+++ scheme/src/main/java/tech/ydb/scheme/description/ListDirectoryResult.java
@@ -1,4 +1,4 @@
-package tech.ydb.table.description;
+package tech.ydb.scheme.description;
 
 import java.util.List;
